requirements.txt
streamlit
pydantic
azure-ai-documentintelligence
llama-index-core
llama-parse
llama-index-readers-file
llama-extract
python-dotenv
streamlit-pdf-viewer